Separated reducers into own files
